tech debit: custom hooks for seperation of logic and html
